+++ src/features/Cart/hooks/useSelectedCart.ts
@@ -1,0 +1,6 @@
+import { useCartContext } from '../context/CartProvider';
+
+export const useSelectedCart = () => {
+  const { cartItems } = useCartContext();
+  return cartItems.filter((item) => item.isChecked);
+};
